requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
docutils==0.20.1
fonttools==4.39.4
fpdf==1.7.2
idna==3.4
Kivy==2.2.0
Kivy-examples==2.2.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
plyer==2.1.0
Pygments==2.15.1
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.3
